MSAPI-74 | Change Status Code from string to integer (#27)

+++ SignNow.Net.Test/UnitTests/Exceptions/SignNowExceptionTest.cs
@@ -1,6 +1,5 @@
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using SignNow.Net.Exceptions;
-using System;
 using System.Net;
 
 namespace UnitTests
@@ -26,7 +25,7 @@ namespace UnitTests
             Assert.AreEqual(testMessage, ex.Message, $"Exception {ex.Message.GetType().ToString()} should have '{testMessage}'");
             Assert.AreEqual(HttpStatusCode.BadRequest, ex.HttpStatusCode, $"Exception HttpStatusCode should be '{HttpStatusCode.BadRequest.ToString()}'");
             Assert.IsTrue(ex.Data.Contains("HttpStatusCode"));
-            Assert.AreEqual(HttpStatusCode.BadRequest, ex.Data["HttpStatusCode"]);
+            Assert.AreEqual((int)HttpStatusCode.BadRequest, ex.Data["HttpStatusCode"]);
         }
 
         [TestMethod]
